chore(inputs.docker_logs): Cleanup plugin code
