+++ plugins/inputs/docker_log/docker_log.go
@@ -2,20 +2,14 @@
 package docker_log
 
 import (
-	"bufio"
-	"bytes"
 	"context"
-	"crypto/tls"
 	_ "embed"
 	"errors"
 	"fmt"
-	"io"
-	"strings"
+	"net/http"
 	"sync"
 	"time"
-	"unicode"
 
-	"github.com/moby/moby/api/pkg/stdcopy"
 	"github.com/moby/moby/api/types/container"
 	"github.com/moby/moby/client"
 
@@ -44,10 +38,7 @@ type DockerLogs struct {
 
 	common_tls.ClientConfig
 
-	newEnvClient func() (dockerClient, error)
-	newClient    func(string, *tls.Config) (dockerClient, error)
-
-	client          dockerClient
+	client          *client.Client
 	labelFilter     filter.Filter
 	containerFilter filter.Filter
 	stateFilter     filter.Filter
@@ -66,36 +57,65 @@ func (*DockerLogs) SampleConfig() string {
 }
 
 func (d *DockerLogs) Init() error {
-	if d.Endpoint == "ENV" {
-		c, err := d.newEnvClient()
+	if d.Endpoint == "" {
+		d.Endpoint = "unix:///var/run/docker.sock"
+	}
+
+	switch d.Endpoint {
+	case "ENV":
+		c, err := client.New(client.FromEnv)
 		if err != nil {
 			return fmt.Errorf("creating client from environment failed: %w", err)
 		}
 		d.client = c
-	} else {
+	default:
 		tlsConfig, err := d.ClientConfig.TLSConfig()
 		if err != nil {
 			return fmt.Errorf("creating TLS configuration failed: %w", err)
 		}
-		c, err := d.newClient(d.Endpoint, tlsConfig)
+		transport := &http.Transport{
+			TLSClientConfig: tlsConfig,
+		}
+		c, err := client.New(
+			client.WithHTTPHeaders(map[string]string{"User-Agent": "engine-api-cli-1.0"}),
+			client.WithHTTPClient(&http.Client{Transport: transport}),
+			client.WithHost(d.Endpoint),
+		)
 		if err != nil {
 			return fmt.Errorf("creating client failed: %w", err)
 		}
 		d.client = c
 	}
 
-	// Create filters
-	if err := d.createLabelFilters(); err != nil {
+	// Create label filter
+	labelFilter, err := filter.NewIncludeExcludeFilter(d.LabelInclude, d.LabelExclude)
+	if err != nil {
 		return fmt.Errorf("creating label filter failed: %w", err)
 	}
-	if err := d.createContainerFilters(); err != nil {
+	d.labelFilter = labelFilter
+
+	// Create container filer
+	containerFilter, err := filter.NewIncludeExcludeFilter(d.ContainerInclude, d.ContainerExclude)
+	if err != nil {
 		return fmt.Errorf("creating container filter failed: %w", err)
 	}
-	if err := d.createContainerStateFilters(); err != nil {
+	d.containerFilter = containerFilter
+
+	// Create container state filter
+	if len(d.ContainerStateInclude) == 0 && len(d.ContainerStateExclude) == 0 {
+		d.ContainerStateInclude = []string{"running"}
+	}
+	stateFilter, err := filter.NewIncludeExcludeFilter(d.ContainerStateInclude, d.ContainerStateExclude)
+	if err != nil {
 		return fmt.Errorf("creating container state filter failed: %w", err)
 	}
+	d.stateFilter = stateFilter
 
+	// Allocate maps
 	d.lastRecord = make(map[string]time.Time)
+	d.mu.Lock()
+	d.containerList = make(map[string]context.CancelFunc)
+	d.mu.Unlock()
 
 	return nil
 }
@@ -136,18 +156,21 @@ func (d *DockerLogs) Gather(acc telegraf.Accumulator) error {
 
 	ctx, cancel := context.WithTimeout(ctx, time.Duration(d.Timeout))
 	defer cancel()
+
 	containers, err := d.client.ContainerList(ctx, client.ContainerListOptions{})
 	if err != nil {
-		return err
+		return fmt.Errorf("listing containers failed: %w", err)
 	}
 
-	for _, cntnr := range containers {
-		if d.containerInContainerList(cntnr.ID) {
+	for _, cntnr := range containers.Items {
+		d.mu.Lock()
+		_, found := d.containerList[cntnr.ID]
+		d.mu.Unlock()
+		if found {
 			continue
 		}
 
 		containerName := parseContainerName(cntnr.Names)
-
 		if containerName == "" {
 			continue
 		}
@@ -161,17 +184,24 @@ func (d *DockerLogs) Gather(acc telegraf.Accumulator) error {
 		}
 
 		ctx, cancel := context.WithCancel(context.Background())
-		d.addToContainerList(cntnr.ID, cancel)
+		d.mu.Lock()
+		d.containerList[cntnr.ID] = cancel
+		d.mu.Unlock()
 
 		// Start a new goroutine for every new container that has logs to collect
 		d.wg.Add(1)
 		go func(container container.Summary) {
 			defer d.wg.Done()
-			defer d.removeFromContainerList(container.ID)
+			defer func() {
+				d.mu.Lock()
+				defer d.mu.Unlock()
+				delete(d.containerList, container.ID)
+			}()
 
-			err = d.tailContainerLogs(ctx, acc, container, containerName)
-			if err != nil && !errors.Is(err, context.Canceled) {
-				acc.AddError(err)
+			if err := d.tailContainerLogs(ctx, acc, container, containerName); err != nil {
+				if !errors.Is(err, context.Canceled) {
+					acc.AddError(err)
+				}
 			}
 		}(cntnr)
 	}
@@ -181,25 +211,6 @@ func (d *DockerLogs) Gather(acc telegraf.Accumulator) error {
 func (d *DockerLogs) Stop() {
 	d.cancelTails()
 	d.wg.Wait()
-}
-
-func (d *DockerLogs) addToContainerList(containerID string, cancel context.CancelFunc) {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-	d.containerList[containerID] = cancel
-}
-
-func (d *DockerLogs) removeFromContainerList(containerID string) {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-	delete(d.containerList, containerID)
-}
-
-func (d *DockerLogs) containerInContainerList(containerID string) bool {
-	d.mu.Lock()
-	defer d.mu.Unlock()
-	_, ok := d.containerList[containerID]
-	return ok
 }
 
 func (d *DockerLogs) cancelTails() {
@@ -213,22 +224,24 @@ func (d *DockerLogs) cancelTails() {
 func (d *DockerLogs) hasTTY(ctx context.Context, cntnr container.Summary) (bool, error) {
 	ctx, cancel := context.WithTimeout(ctx, time.Duration(d.Timeout))
 	defer cancel()
-	c, err := d.client.ContainerInspect(ctx, cntnr.ID)
+
+	result, err := d.client.ContainerInspect(ctx, cntnr.ID, client.ContainerInspectOptions{})
 	if err != nil {
-		return false, err
+		return false, fmt.Errorf("inspecting container %q failed: %w", cntnr.ID, err)
 	}
-	return c.Config.Tty, nil
+
+	return result.Container.Config.Tty, nil
 }
 
 func (d *DockerLogs) tailContainerLogs(
 	ctx context.Context,
 	acc telegraf.Accumulator,
 	cntnr container.Summary,
-	containerName string,
+	name string,
 ) error {
 	imageName, imageVersion := docker.ParseImage(cntnr.Image)
 	tags := map[string]string{
-		"container_name":    containerName,
+		"container_name":    name,
 		"container_image":   imageName,
 		"container_version": imageVersion,
 	}
@@ -297,177 +310,10 @@ func (d *DockerLogs) tailContainerLogs(
 	return nil
 }
 
-// Parse container name
-func parseContainerName(containerNames []string) string {
-	for _, name := range containerNames {
-		trimmedName := strings.TrimPrefix(name, "/")
-		if !strings.Contains(trimmedName, "/") {
-			return trimmedName
-		}
-	}
-
-	return ""
-}
-
-func parseLine(line []byte) (time.Time, string, error) {
-	parts := bytes.SplitN(line, []byte(" "), 2)
-
-	if len(parts) == 1 {
-		parts = append(parts, []byte(""))
-	}
-
-	tsString := string(parts[0])
-
-	// Keep any leading space, but remove whitespace from end of line.
-	// This preserves space in, for example, stacktraces, while removing
-	// annoying end of line characters and is similar to how other logging
-	// plugins such as syslog behave.
-	message := bytes.TrimRightFunc(parts[1], unicode.IsSpace)
-
-	ts, err := time.Parse(time.RFC3339Nano, tsString)
-	if err != nil {
-		return time.Time{}, "", fmt.Errorf("error parsing timestamp %q: %w", tsString, err)
-	}
-
-	return ts, string(message), nil
-}
-
-func tailStream(
-	acc telegraf.Accumulator,
-	baseTags map[string]string,
-	containerID string,
-	reader io.ReadCloser,
-	stream string,
-) (time.Time, error) {
-	defer reader.Close()
-
-	tags := make(map[string]string, len(baseTags)+1)
-	for k, v := range baseTags {
-		tags[k] = v
-	}
-	tags["stream"] = stream
-
-	r := bufio.NewReaderSize(reader, 64*1024)
-
-	var lastTS time.Time
-	for {
-		line, err := r.ReadBytes('\n')
-
-		if len(line) != 0 {
-			ts, message, err := parseLine(line)
-			if err != nil {
-				acc.AddError(err)
-			} else {
-				acc.AddFields("docker_log", map[string]interface{}{
-					"container_id": containerID,
-					"message":      message,
-				}, tags, ts)
-			}
-
-			// Store the last processed timestamp
-			if ts.After(lastTS) {
-				lastTS = ts
-			}
-		}
-
-		if err != nil {
-			if err == io.EOF {
-				return lastTS, nil
-			}
-			return time.Time{}, err
-		}
-	}
-}
-
-func tailMultiplexed(acc telegraf.Accumulator, tags map[string]string, containerID string, src io.ReadCloser) (time.Time, error) {
-	outReader, outWriter := io.Pipe()
-	errReader, errWriter := io.Pipe()
-
-	var tsStdout, tsStderr time.Time
-	var wg sync.WaitGroup
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
-		var err error
-		tsStdout, err = tailStream(acc, tags, containerID, outReader, "stdout")
-		if err != nil {
-			acc.AddError(err)
-		}
-	}()
-
-	wg.Add(1)
-	go func() {
-		defer wg.Done()
-		var err error
-		tsStderr, err = tailStream(acc, tags, containerID, errReader, "stderr")
-		if err != nil {
-			acc.AddError(err)
-		}
-	}()
-
-	_, err := stdcopy.StdCopy(outWriter, errWriter, src)
-
-	// Ignore the returned errors as we cannot do anything if the closing fails
-	_ = outWriter.Close()
-	_ = errWriter.Close()
-	_ = src.Close()
-	wg.Wait()
-
-	if err != nil {
-		return time.Time{}, err
-	}
-	if tsStdout.After(tsStderr) {
-		return tsStdout, nil
-	}
-	return tsStderr, nil
-}
-
-// Following few functions have been inherited from telegraf docker input plugin
-func (d *DockerLogs) createContainerFilters() error {
-	containerFilter, err := filter.NewIncludeExcludeFilter(d.ContainerInclude, d.ContainerExclude)
-	if err != nil {
-		return err
-	}
-	d.containerFilter = containerFilter
-	return nil
-}
-
-func (d *DockerLogs) createLabelFilters() error {
-	labelFilter, err := filter.NewIncludeExcludeFilter(d.LabelInclude, d.LabelExclude)
-	if err != nil {
-		return err
-	}
-	d.labelFilter = labelFilter
-	return nil
-}
-
-func (d *DockerLogs) createContainerStateFilters() error {
-	if len(d.ContainerStateInclude) == 0 && len(d.ContainerStateExclude) == 0 {
-		d.ContainerStateInclude = []string{"running"}
-	}
-	stateFilter, err := filter.NewIncludeExcludeFilter(d.ContainerStateInclude, d.ContainerStateExclude)
-	if err != nil {
-		return err
-	}
-	d.stateFilter = stateFilter
-	return nil
-}
-
-func hostnameFromID(id string) string {
-	if len(id) > 12 {
-		return id[0:12]
-	}
-	return id
-}
-
 func init() {
 	inputs.Add("docker_log", func() telegraf.Input {
 		return &DockerLogs{
-			Timeout:       config.Duration(time.Second * 5),
-			Endpoint:      "unix:///var/run/docker.sock",
-			newEnvClient:  newEnvClient,
-			newClient:     newClient,
-			containerList: make(map[string]context.CancelFunc),
+			Timeout: config.Duration(time.Second * 5),
 		}
 	})
 }
