+++ plugins/inputs/docker_log/docker_log_test.go
@@ -123,7 +123,6 @@ func TestGather(t *testing.T) {
 				IncludeSourceTag: true,
 				ClientConfig:     common_tls.ClientConfig{InsecureSkipVerify: true}, // Required as the test server has only a self-signed cert
 				Timeout:          config.Duration(time.Second * 5),
-				newClient:        newClient,
 				containerList:    make(map[string]context.CancelFunc),
 			}
 			require.NoError(t, plugin.Init())
